+++ Makefile
@@ -4,7 +4,7 @@ LIBS    = libft/libft.a
 INCLUDE = -I./include -I./libft/include -I $(brew --prefix readline)/include
 NAME    = minishell
 SRCDIR  = src
-SRCS    = src/lexer/lexer_helper.c src/lexer/lexer.c src/utils/error.c src/parser/ast.c src/parser/parser.c src/main.c src/builtin/mini_echo.c src/builtin/mini_pwd.c
+SRCS    = src/lexer/lexer_helper.c src/lexer/lexer.c src/utils/error.c src/parser/ast.c src/parser/parser.c src/main.c src/executor/executor.c src/builtin/mini_echo.c src/builtin/mini_pwd.c
 OBJDIR  = obj
 OBJS    = $(subst $(SRCDIR), $(OBJDIR), $(SRCS:.c=.o))
 DEPENDS = $(OBJS:.o=.d)
