+++ include/executor.h
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   executor.h                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/03 18:52:30 by tkuramot          #+#    #+#             */
+/*   Updated: 2023/09/03 18:53:16 by tkuramot         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef EXECUTOR_H
+# define EXECUTOR_H
+
+# include "parser.h"
+
+void	execute_cmd(t_ast *ast);
+
+#endif
